Moved database configuration
