+++ ebms-async/src/main/kotlin/no/nav/emottak/ebms/async/App.kt
@@ -31,15 +31,15 @@ import no.nav.emottak.ebms.async.configuration.config
 import no.nav.emottak.ebms.async.kafka.consumer.startPayloadReceiver
 import no.nav.emottak.ebms.async.kafka.consumer.startSignalReceiver
 import no.nav.emottak.ebms.async.kafka.producer.EbmsMessageProducer
+import no.nav.emottak.ebms.async.persistence.Database
+import no.nav.emottak.ebms.async.persistence.ebmsDbConfig
+import no.nav.emottak.ebms.async.persistence.ebmsMigrationConfig
+import no.nav.emottak.ebms.async.persistence.repository.EbmsMessageDetailsRepository
+import no.nav.emottak.ebms.async.persistence.repository.PayloadRepository
 import no.nav.emottak.ebms.async.processing.PayloadMessageProcessor
 import no.nav.emottak.ebms.async.processing.PayloadMessageResponder
 import no.nav.emottak.ebms.async.processing.SignalProcessor
 import no.nav.emottak.ebms.defaultHttpClient
-import no.nav.emottak.ebms.persistence.Database
-import no.nav.emottak.ebms.persistence.ebmsDbConfig
-import no.nav.emottak.ebms.persistence.ebmsMigrationConfig
-import no.nav.emottak.ebms.persistence.repository.EbmsMessageDetailsRepository
-import no.nav.emottak.ebms.persistence.repository.PayloadRepository
 import no.nav.emottak.ebms.processing.ProcessingService
 import no.nav.emottak.ebms.registerHealthEndpoints
 import no.nav.emottak.ebms.registerNavCheckStatus
