+++ ebms-async/src/main/kotlin/no/nav/emottak/ebms/async/Routes.kt
@@ -4,7 +4,7 @@ import io.ktor.http.HttpStatusCode
 import io.ktor.server.response.respond
 import io.ktor.server.routing.Route
 import io.ktor.server.routing.get
-import no.nav.emottak.ebms.persistence.repository.PayloadRepository
+import no.nav.emottak.ebms.async.persistence.repository.PayloadRepository
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
