+++ ebms-async/src/main/kotlin/no/nav/emottak/ebms/async/persistence/repository/EventsRepository.kt
@@ -1,15 +1,18 @@
-package no.nav.emottak.ebms.persistence.repository
+package no.nav.emottak.ebms.async.persistence.repository
 
-import no.nav.emottak.ebms.persistence.Database
-import no.nav.emottak.ebms.persistence.table.EventsTable
-import no.nav.emottak.ebms.persistence.table.EventsTable.contentId
-import no.nav.emottak.ebms.persistence.table.EventsTable.createdAt
-import no.nav.emottak.ebms.persistence.table.EventsTable.eventId
-import no.nav.emottak.ebms.persistence.table.EventsTable.eventMessage
-import no.nav.emottak.ebms.persistence.table.EventsTable.juridiskLoggId
-import no.nav.emottak.ebms.persistence.table.EventsTable.messageId
-import no.nav.emottak.ebms.persistence.table.EventsTable.requestId
+import no.nav.emottak.ebms.async.persistence.Database
+import no.nav.emottak.ebms.async.persistence.table.EventsTable
+import no.nav.emottak.ebms.async.persistence.table.EventsTable.contentId
+import no.nav.emottak.ebms.async.persistence.table.EventsTable.createdAt
+import no.nav.emottak.ebms.async.persistence.table.EventsTable.eventId
+import no.nav.emottak.ebms.async.persistence.table.EventsTable.eventMessage
+import no.nav.emottak.ebms.async.persistence.table.EventsTable.juridiskLoggId
+import no.nav.emottak.ebms.async.persistence.table.EventsTable.messageId
+import no.nav.emottak.ebms.async.persistence.table.EventsTable.requestId
+import no.nav.emottak.ebms.util.marker
+import no.nav.emottak.message.model.EbmsMessage
 import no.nav.emottak.message.model.Event
+import no.nav.emottak.message.model.log
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.jetbrains.exposed.sql.upsert
 import kotlin.uuid.ExperimentalUuidApi
@@ -59,4 +62,17 @@ class EventsRepository(private val database: Database) {
         }
         return event
     }
+}
+
+@OptIn(ExperimentalUuidApi::class)
+fun EbmsMessage.saveEvent(message: String, eventsRepository: EventsRepository) {
+    log.info(this.marker(), message)
+    eventsRepository.updateOrInsert(
+        Event(
+            eventId = Uuid.random(),
+            requestId = Uuid.parse(this.requestId),
+            messageId = this.messageId,
+            eventMessage = message
+        )
+    )
 }
