+++ ebms-async/src/main/kotlin/no/nav/emottak/ebms/async/processing/SignalProcessor.kt
@@ -3,7 +3,7 @@ package no.nav.emottak.ebms.async.processing
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import no.nav.emottak.ebms.async.log
-import no.nav.emottak.ebms.persistence.repository.EbmsMessageDetailsRepository
+import no.nav.emottak.ebms.async.persistence.repository.EbmsMessageDetailsRepository
 import no.nav.emottak.ebms.util.marker
 import no.nav.emottak.ebms.validation.DokumentValidator
 import no.nav.emottak.message.model.Acknowledgment
