+++ ebms-async/src/main/resources/db/migration/V3__replace_ebms_message_with_ebms_message_details.sql
