+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/App.kt
@@ -74,7 +74,6 @@ fun Application.ebmsProviderModule(
     installMicrometerRegistry(appMicrometerRegistry)
     installRequestTimerPlugin()
     installContentNegotiation()
-    installAuthentication()
 
     routing {
         registerRootEndpoint()
