+++ ebms-provider/src/main/kotlin/no/nav/emottak/ebms/Plugins.kt
@@ -4,7 +4,6 @@ import io.ktor.serialization.kotlinx.json.json
 import io.ktor.server.application.Application
 import io.ktor.server.application.createRouteScopedPlugin
 import io.ktor.server.application.install
-import io.ktor.server.auth.Authentication
 import io.ktor.server.metrics.micrometer.MicrometerMetrics
 import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.server.request.uri
@@ -12,7 +11,6 @@ import io.ktor.util.logging.KtorSimpleLogger
 import io.micrometer.prometheusmetrics.PrometheusMeterRegistry
 import net.logstash.logback.marker.Markers
 import no.nav.emottak.constants.SMTPHeaders
-import no.nav.security.token.support.v3.tokenValidationSupport
 import java.time.Duration
 import java.time.Instant
 import kotlin.time.toKotlinDuration
@@ -26,12 +24,6 @@ internal fun Application.installMicrometerRegistry(appMicrometerRegistry: Promet
 internal fun Application.installContentNegotiation() {
     install(ContentNegotiation) {
         json()
-    }
-}
-
-internal fun Application.installAuthentication() {
-    install(Authentication) {
-        tokenValidationSupport(AZURE_AD_AUTH, AuthConfig.getTokenSupportConfig())
     }
 }
